requirements.txt
django
django-jsonfield
django_countries
markdown
requests
pygal
pytz
python-bitcoinlib
stripe

